+++ history.md
@@ -1,6 +1,7 @@
 ## Edge
 
 * Fix: String#translation_for('xx') should not return an array of all translations. #88
+* Fix: Term#translation_for('xx') should not return an array of all term translations.
 
 ## Version 2.0.2 / 2012-06-06